cloud-run-ai-agents/requirements.txt
# Google ADK and AI dependencies
google-adk>=0.1.0
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.0
google-genai>=0.1.0

# Google Cloud dependencies
google-cloud-pubsub>=2.20.0
google-cloud-storage>=2.16.0
google-cloud-secret-manager>=2.18.0
google-auth>=2.28.0

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# Utilities
python-dotenv>=1.0.1
requests>=2.31.0
Pillow>=10.2.0
aiohttp>=3.9.3
